fb_digest/requirements.txt
playwright>=1.43.0
pyyaml>=6.0
